backend/requirements.txt
fastapi
uvicorn[standard]
mangum
boto3
httpx
pydantic
pydantic-settings
python-multipart
stripe
PyJWT
openai>=1.40.0

